perf & refactor: implement useUpdateRedeemTransactions and useUpdateIssueTransactions
