+++ src/services/use-update-redeem-transactions.ts
@@ -1,0 +1,76 @@
+
+// ray test touch <
+import {
+  useSelector,
+  useDispatch
+} from 'react-redux';
+import * as polkabtcStats from '@interlay/polkabtc-stats';
+import {
+  RedeemColumns,
+  BtcNetworkName
+} from '@interlay/polkabtc-stats';
+
+import useInterval from 'utils/hooks/use-interval';
+import { updateAllRedeemRequestsAction } from 'common/actions/redeem.actions';
+import { statsToUIRedeemRequest } from 'common/utils/utils';
+import { RedeemRequest } from 'common/types/redeem.types';
+import { StoreType } from 'common/types/util.types';
+import * as constants from '../constants';
+
+const useUpdateRedeemTransactions = (
+  page: number = 0,
+  limit: number = 15,
+  delay: number = 10000
+) => {
+  const {
+    address,
+    bitcoinHeight,
+    polkaBtcLoaded
+  } = useSelector((state: StoreType) => state.general);
+  const dispatch = useDispatch();
+
+  const isRunning = address && polkaBtcLoaded;
+
+  useInterval(async () => {
+    try {
+      // Temporary declaration pending refactor decision
+      const stats = new polkabtcStats.StatsApi(new polkabtcStats.Configuration({ basePath: constants.STATS_URL }));
+
+      const [
+        parachainHeight,
+        redeemPeriod,
+        requiredBtcConfirmations
+      ] = await Promise.all([
+        window.polkaBTC.system.getCurrentBlockNumber(), // TODO: should avoid as it's called for issue
+        window.polkaBTC.redeem.getRedeemPeriod(),
+        window.polkaBTC.btcRelay.getStableBitcoinConfirmations() // TODO: should avoid as it's called for issue
+      ]);
+
+      const databaseRequests: RedeemRequest[] = (
+        await stats.getFilteredRedeems(
+          page, // Page 0 (i.e. first page)
+          limit, // 15 per page (i.e. fetch 15 latest requests)
+          undefined, // Default sorting
+          undefined, // Default sort order
+          constants.BITCOIN_NETWORK as BtcNetworkName,
+          [{ column: RedeemColumns.Requester, value: address }] // Filter by requester == address
+        )
+      ).data.map(statsRedeem =>
+        statsToUIRedeemRequest(
+          statsRedeem,
+          bitcoinHeight,
+          parachainHeight,
+          redeemPeriod,
+          requiredBtcConfirmations
+        )
+      );
+
+      dispatch(updateAllRedeemRequestsAction(address, databaseRequests));
+    } catch (error) {
+      console.log('[useUpdateRedeemTransactions useInterval] error.message => ', error.message);
+    }
+  }, isRunning ? delay : null, true);
+};
+
+export default useUpdateRedeemTransactions;
+// ray test touch >
